# web/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Partner Api",
        "version": "1.0.0"
    },
    "host": "api5.audi.dev.agence-one.net",
    "basePath": "/partner/dev/",
    "schemes": [
        "http"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/types": {
            "get": {
                "tags": [
                    "Partner"
                ],
                "summary": "Get partner type",
                "responses": {
                    "200": {
                        "description": "Types contract partner",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Context"
                                },
                                {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/{id}": {
            "get": {
                "tags": [
                    "Partner"
                ],
                "summary": "Get a Partner from id",
                "operationId": "getPartnerById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "partnerId",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Target partner",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Context"
                                },
                                {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/PartnerComplete"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Partner"
                ],
                "summary": "update partner from userId",
                "operationId": "putPartnerById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "partnerId",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "partner",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Partner"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "the updated partner",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Context"
                                },
                                {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/PartnerComplete"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "updating error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Partner"
                ],
                "summary": "remove partner from id",
                "operationId": "removePartnerById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "partnerId",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "the partner",
                        "schema": {
                            "$ref": "#/definitions/Success"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Partner"
                ],
                "summary": "patch partner from id",
                "operationId": "patchPartnerById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "partnerId",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "partner",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Partner"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "the updated partner",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Context"
                                },
                                {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/PartnerComplete"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "updating error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/": {
            "get": {
                "tags": [
                    "Partner"
                ],
                "summary": "get partners",
                "operationId": "getPartners",
                "parameters": [
                    {
                        "name": "registryUserId",
                        "in": "query",
                        "description": "registryUserId",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "myaudiUserId",
                        "in": "query",
                        "description": "myaudiUserId",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page number",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "max items per page",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Target Partners",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Context"
                                },
                                {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/PartnerComplete"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Partner"
                ],
                "summary": "create partner",
                "operationId": "createPartner",
                "parameters": [
                    {
                        "name": "partner",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Partner"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "the created partner",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Context"
                                },
                                {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/PartnerComplete"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
        "Context": {
            "properties": {
                "context": {
                    "type": "string"
                }
            },
            "type": "object"
        },
...
        "Success": {
            "properties": {
                "data": {
                    "properties": {
                        "message": {
                            "type": "string"
                        }
                    },
                    "type": "object"
                }
            },
            "type": "object",
            "allOf": [
                {
                    "$ref": "#/definitions/Context"
                }
            ]
        },
        "Error": {
            "required": [
                "errors"
            ],
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "properties": {
                            "message": {
                                "type": "string"
                            },
                            "code": {
                                "type": "integer"
                            },
                            "type": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "type": "object",
            "allOf": [
                {
                    "$ref": "#/definitions/Context"
                }
            ]
        },
        "IdableEntity": {
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "LegacyEntity": {
            "required": [
                "legacyId"
            ],
            "properties": {
                "legacyId": {
                    "type": "integer"
                }
            }
        },
        "TimestampableEntity": {
            "required": [
                "createdAt",
                "updatedAt"
            ],
            "properties": {
                "createdAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "updatedAt": {
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
...
        "Partner": {
            "allOf": [
                {
                    "required": [
                        "type"
                    ],
                    "properties": {
                        "type": {
                            "type": "string",
                            "enum": [
                                "sales",
                                "aftersales"
                            ]
                        },
                        "contractNumber": {
                            "type": "string"
                        },
                        "commercialName": {
                            "type": "string"
                        },
                        "kvpsNumber": {
                            "type": "string"
                        },
                        "webSite": {
                            "type": "string"
                        },
                        "isPartnerR8": {
                            "type": "boolean"
                        },
                        "isTwinService": {
                            "type": "boolean"
                        },
                        "isPartnerPlus": {
                            "type": "boolean"
                        },
                        "isOccPlus": {
                            "type": "boolean"
                        },
                        "occPlusContractNumber": {
                            "type": "string"
                        },
                        "isEtron": {
                            "type": "boolean"
                        },
                        "group": {
                            "type": "integer"
                        },
                        "siteType": {
                            "type": "string",
                            "enum": [
                                "principal",
                                "secondary"
                            ]
                        },
                        "category": {
                            "type": "string",
                            "enum": [
                                "A",
                                "A*"
                            ]
                        },
                        "representationType": {
                            "type": "string",
                            "enum": [
                                "E1",
                                "E2",
                                "Y",
                                "W",
                                "G"
                            ]
                        },
                        "prestigeType": {
                            "type": "string",
                            "enum": [
                                "exclusive",
                                "specialized"
                            ]
                        },
                        "dealersMeeting": {
                            "type": "boolean"
                        },
                        "brandDays": {
                            "type": "boolean"
                        },
                        "rent": {
                            "type": "boolean"
                        },
                        "extraHour": {
                            "type": "boolean"
                        },
                        "ferMembership": {
                            "type": "boolean"
                        },
                        "onlineQuotation": {
                            "type": "boolean"
                        },
                        "amexPayment": {
                            "type": "boolean"
                        },
                        "isDigitAll": {
                            "type": "boolean"
                        },
                        "digitAllId": {
                            "type": "string"
                        },
                        "isV12": {
                            "type": "boolean"
                        },
                        "v12Id": {
                            "type": "string"
                        },
                        "sellingVolume": {
                            "type": "integer"
                        }
                    }
                },
                {
                    "$ref": "#/definitions/LegacyEntity"
                }
            ]
        },
        "PartnerComplete": {
            "allOf": [
                {
                    "$ref": "#/definitions/IdableEntity"
                },
                {
                    "$ref": "#/definitions/Partner"
                },
                {
                    "$ref": "#/definitions/TimestampableEntity"
                }
            ]
        },
...
    },
    "securityDefinitions": {
        "bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}